+++ FindPanel/OgreTextFindComponents/OgreTextFindLeaf.h
@@ -19,7 +19,7 @@
 @interface OgreTextFindLeaf : NSObject <OgreTextFindComponent>
 {
     OgreTextFindBranch      *_parent;
-    int                     _index;
+    NSInteger               _index;
     BOOL                    _isParentRetained;
     
     BOOL                    _isTerminal;
